+++ sagaart/news/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from .models import NewsModel
+
+
+@admin.register(NewsModel)
+class NewsModelAdmin(admin.ModelAdmin):
+    list_display = ('id', 'name', 'text', 'is_active', 'date_pub')
