+++ sagaart/sagaart/settings.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = [
     'corsheaders',
     'api',
     'users',
+    'news',
     'django_filters',
     'rest_framework',
     'drf_spectacular',
